用户管理员信息展示接口 已完成 用户家长信息展示接口与添加接口 已完成 用户教练信息展示接口 已完成

+++ src/main/java/com/yjy/model/Parent.java
@@ -43,9 +43,9 @@ public class Parent {
     /**
      * 注册时间
      */
-    private BigInteger parentRegtime;
+    private Long parentRegtime;
     /**
      * 修改时间
      */
-    private BigInteger updateTime;
+    private Long updateTime;
 }
